fix(deprize): resolve Suspense hydration error and restore card navigation
Load the list client-only to avoid Suspense hydration races that blocked
clicks via the Next.js error overlay. Use real Links for card navigation
and make team odds rows full-width Bet buttons.
